fix: ADD temp min and max features to features df
As the window size is 1 second long, and the frequency of the TEMP
signal is 4 Hz. The statistics computed in each window is quite similar
which involves to produce equal columns. The drop_duplicate() function
catches this equality and drops the columns. The commit adds manually
this features to the features dataframe